Trees restructure
