+++ Trees, Binary Trees, & Binary Search Trees/BSTInsertionAndRemoval/Delete.java
@@ -1,20 +1,4 @@
-class InsertIntoABST {
-  public TreeNode insertIntoBST(TreeNode root, int val) {
-    if (root == null) {
-      return new TreeNode(val);
-    }
-
-    if (root.val > val) {
-      root.left = insertIntoBST(root.left, val);
-    } else {
-      root.right = insertIntoBST(root.right, val);
-    }
-
-    return root;
-  }
-}
-
-class DeleteFromABST {
+class Solution {
   public TreeNode deleteNode(TreeNode root, int key) {
     if (root == null) {
       return null;
