+++ Trees, Binary Trees, & Binary Search Trees/LevelOrderTraversal/LevelOrderTraversal.java
@@ -1,86 +1,55 @@
-public List<List<Integer>> levelOrder(TreeNode root) {
-
-  if (root == null) {
-    return new ArrayList<>();
-  }
-
-  /*
-    The list to hold the level by level layers of the tree
-  */
-  List<List<Integer>> levelsList = new ArrayList<List<Integer>>();
-
-  /*
-    Our queue to enforce a traversal in a breadth-first manner
-  */
-  Queue<TreeNode> queue = new LinkedList<>();
-
-  /*
-    Add the root node to start the breadth first search
-  */
-  queue.offer(root);
-
-  /*
-    We will continue the breadth first search as long as
-    there are nodes to process.
-    
-    We do not need a hashtable to track nodes already
-    visited since a tree is an acyclic connected graph.
-
-    Each iteration will process a layer of the tree.
-  */
-  while (!queue.isEmpty()) {
-
-    /*
-      The list to hold all the node's values in the
-      layer we are about to process
-    */
-    List<Integer> currentLayer = new ArrayList<>();
-
-    /*
-      We will get the size of the layer in the queue
-      that we need to process so that we know how
-      many nodes our for loop needs to process
-    */
-    int layerSize = queue.size();
-
-    for (int i = 0; i < layerSize; i++) {
-
-      /*
-        Remove the nodes so that we can process it
-      */
-      TreeNode currentNode = queue.poll();
-
-      /*
-        Add the node's value to the current layer list
-      */
-      currentLayer.add(currentNode.val);
-
-      /*
-        If this node has a left add it to be
-        processed for the next layer
-      */
-      if (currentNode.left != null) {
-        queue.offer(currentNode.left);
-      }
-
-      /*
-        If this node has a right add it to be
-        processed for the next layer
-      */
-      if (currentNode.right != null) {
-        queue.offer(currentNode.right);
-      }
-
+class Solution {
+  public List<List<Integer>> levelOrderTraversal(TreeNode root) {
+    if (root == null) {
+      return new ArrayList<>();
     }
 
+    List<List<Integer>> levelsList = new ArrayList<List<Integer>>();
+
+    Queue<TreeNode> queue = new LinkedList<>();
+    queue.offer(root);
+
     /*
-      Notice how currentLayer is a new list on each iteration
-      of the while loop. We do not need to worry about deep
-      copying anything here.
+      We will continue the breadth first search as long as
+      there are nodes to process.
+      
+      We do not need a hashtable to track nodes already
+      visited since a tree is an acyclic connected graph.
+
+      Each iteration will process a layer of the tree.
     */
-    levelsList.add(currentLayer);
+    while (!queue.isEmpty()) {
+      List<Integer> currentLayer = new ArrayList<>();
 
+      /*
+        We will get the size of the layer in the queue
+        that we need to process so that we know how
+        many nodes our for loop needs to process
+      */
+      int layerSize = queue.size();
+
+      for (int i = 0; i < layerSize; i++) {
+        TreeNode currentNode = queue.poll();
+
+        currentLayer.add(currentNode.val);
+
+        if (currentNode.left != null) {
+          queue.offer(currentNode.left);
+        }
+
+        if (currentNode.right != null) {
+          queue.offer(currentNode.right);
+        }
+      }
+
+      /*
+        Notice how currentLayer is a new list on each iteration
+        of the while loop. We do not need to worry about deep
+        copying anything here.
+      */
+      levelsList.add(currentLayer);
+    }
+
+    return levelsList;
   }
-
-  return levelsList;
 }
