+++ Trees, Binary Trees, & Binary Search Trees/MinHeightBst/MinHeightBst.java
@@ -24,5 +24,4 @@ class Solution {
 
     return newNode;
   }
-
 }
