+++ Trees, Binary Trees, & Binary Search Trees/PopulatingLevelPointers/FragmentedVariant.java
@@ -1,25 +1,4 @@
-class PartI {
-  public Node connect(Node root) {
-    if (root == null) {
-      return null;
-    }
-
-    if (root.left != null) {
-      root.left.next = root.right;
-    }
-
-    if (root.right != null && root.next != null) {
-      root.right.next = root.next.left;
-    }
-
-    root.left = connect(root.left);
-    root.right = connect(root.right);
-
-    return root;
-  }
-}
-
-class PartII {
+class Solution {
   public Node connect(Node root) {
     if (root == null) {
       return null;
