+++ Trees, Binary Trees, & Binary Search Trees/SymmetricBinaryTree/SymmetricBinaryTree.java
@@ -19,5 +19,4 @@ class Solution {
 
     return false;
   }
-
 }
